CTRL_SHIFT/requirements.txt
adapter==0.1
anytree==2.8.0
asgiref==3.2.10
automium==0.2.6
automium-web==0.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
colorlog==4.1.0
Django==3.1
dnspython==1.16.0
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
legacy==0.1.6
logger==1.4
MarkupSafe==1.1.1
my-module==1.6.2
mysql-connector==2.2.9
mysql-connector-python==8.0.19
mysql-connector-python-rf==2.2.2
numpy==1.17.4
opencv-python==4.1.2.30
pika==1.1.0
protobuf==3.6.1
pycparser==2.20
pycryptodome==3.9.8
pymsteams==0.1.12
PyMySQL==0.10.0
pyodbc==4.0.30
pytz==2020.1
rabbitmq==0.2.0
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
utils==1.0.1
Werkzeug==1.0.1
